cdk_sample_architecture/requirements.txt
aws-cdk-lib>=2.91.0
constructs>=10.0.0
types-setuptools
